Add preview functionality
